+++ i6_native_ops/monotonic_rnnt/include/gpu_rnnt.h
@@ -64,7 +64,7 @@ class GpuRNNTComputer {
                 printf("  t = %d\n", t);
                 for (int s = 0; s <= 0; s++) {
                     printf("    s = %d\n      ", s);
-                    int denom_idx = denom_start_indices_host[b] + t * (S[b] + 1) + s;
+                    int64_t denom_idx = denom_start_indices_host[b] + t * (S[b] + 1) + s;
                     for (int v = 0; v < V; v++) {
                         printf("%.4f ", cpu_acts[denom_idx * V + v]);
                     }
@@ -81,7 +81,7 @@ class GpuRNNTComputer {
                 printf("  t = %d\n", t);
                 for (int s = 0; s <= 0; s++) {
                     printf("    s = %d\n      ", s);
-                    int denom_idx = denom_start_indices_host[b] + t * (S[b] + 1) + s;
+                    int64_t denom_idx = denom_start_indices_host[b] + t * (S[b] + 1) + s;
                     for (int v = 0; v < V; v++) {
                         printf("%.4f ", exp(cpu_acts[denom_idx * V + v] + cpu_denoms[denom_idx]));
                     }
@@ -201,11 +201,11 @@ class GpuRNNTComputer {
             printf("DEBUG: compute_grad_kernel %.2f ms\n", elapsed.count() * 1000);
 #endif
 #ifdef DEBUG_GRADS
-            std::vector<ProbT> cpu_grads(workspace_manager_.num_denoms() * V);
+            std::vector<ProbT> cpu_grads(static_cast<size_t>(workspace_manager_.num_denoms()) * V);
             cudaMemcpy(cpu_grads.data(), grads, sizeof(ProbT) * cpu_grads.size(), cudaMemcpyDeviceToHost);
 
             printf("gpu grads\n");
-            int grad_idx = 0;
+            int64_t grad_idx = 0;
             for (int b = 0; b < B; ++b) {
                 printf("b = %d\n", b);
                 for (int t = 0; t < T[b]; ++t) {
